Created post and comment system
Both front end and back end have been created.
Localization is also included.

The welcome view displays all posts without comments. It receives a
PostViewModel from the show() method in the PostController.
The post view displays a single post with comments. It also receives a
PostViewModel from the showAll() method in the PostController.

Posts are added using the addPost() method in the PostController.
Comments are added using the addComment method in CommentController.